Address review: consolidate beast3d config tests in test_config.py
Per review on PR #56:
- Move TestBeast3DModelConfig from tests/models/beast3d/test_beast3d_model.py
  to tests/test_config.py, keeping all config-schema tests together.
- Add a get_beast_config_class('beast3d') dispatcher test alongside the erayzer
  one while the config tests are consolidated.

The integration test was already gated with @requires_gsplat_cuda on the
test_integration_basic method. Test-only change; ruff + pyright clean.
