+++ tests/test_config.py
@@ -6,6 +6,7 @@ import pytest
 from pydantic import ValidationError
 
 from beast.config import (
+    Beast3DBeastConfig,
     BeastConfig,
     ERayZerBeastConfig,
     OptimizerConfig,
@@ -13,6 +14,7 @@ from beast.config import (
     get_beast_config_class,
 )
 from beast.io import load_config
+from beast.models.beast3d.beast3d_config import Beast3DModelConfig
 from beast.models.beast_resnet.beast_resnet_config import ResnetModelParams
 from beast.models.beast_vit.beast_vit_config import VitModelParams
 from beast.models.erayzer.erayzer_config import ERayZerOptimizerConfig, ERayZerTrainingConfig
@@ -53,6 +55,8 @@ _MINIMAL_ERAYZER = {
     'optimizer': {'lr': 4e-4},
     'data': {'data_dir': '/path/to/data'},
 }
+
+_MIN_TRANSFORMER = {'d': 768, 'd_head': 64, 'encoder_geom_n_layer': 16}
 
 
 class TestBeastConfig:
@@ -218,11 +222,47 @@ class TestERayZerBeastConfig:
         assert 'beta1' in dumped['optimizer']
 
 
+class TestBeast3DModelConfig:
+    """Test the Beast3DModelConfig schema."""
+
+    def test_valid_config_defaults(self) -> None:
+        cfg = Beast3DModelConfig.model_validate(
+            {'model_class': 'beast3d', 'transformer': _MIN_TRANSFORMER},
+        )
+        assert cfg.model_class == 'beast3d'
+        assert cfg.use_dinov3 is True
+        assert cfg.freeze_dinov3 is True
+        assert cfg.frustum_constraint is True
+        assert cfg.random_background is True
+        assert cfg.mask_loss_weight == 0.1
+
+    def test_wrong_model_class_raises(self) -> None:
+        with pytest.raises(ValidationError):
+            Beast3DModelConfig.model_validate(
+                {'model_class': 'erayzer', 'transformer': _MIN_TRANSFORMER},
+            )
+
+    def test_missing_transformer_raises(self) -> None:
+        with pytest.raises(ValidationError):
+            Beast3DModelConfig.model_validate({'model_class': 'beast3d'})
+
+    def test_inherits_erayzer_fields(self) -> None:
+        # hard_pixelalign and gaussians come from the ERayZer base config
+        cfg = Beast3DModelConfig.model_validate(
+            {'model_class': 'beast3d', 'transformer': _MIN_TRANSFORMER, 'hard_pixelalign': True},
+        )
+        assert cfg.hard_pixelalign is True
+        assert cfg.gaussians.sh_degree == 3
+
+
 class TestGetBeastConfigClass:
     """Test the get_beast_config_class dispatcher."""
 
     def test_erayzer_returns_erayzer_beast_config(self) -> None:
         assert get_beast_config_class('erayzer') is ERayZerBeastConfig
+
+    def test_beast3d_returns_beast3d_beast_config(self) -> None:
+        assert get_beast_config_class('beast3d') is Beast3DBeastConfig
 
     def test_resnet_falls_back_to_beast_config(self) -> None:
         assert get_beast_config_class('resnet') is BeastConfig
